requirements.txt
requests
dload
python-telegram-bot==12.6.1
google-api-python-client>=1.7.11,<1.7.20
google-auth-httplib2>=0.0.3,<0.1.0
google-auth-oauthlib>=0.4.1,<0.10.0
tenacity>=6.0.0
python-magic
wget
asyncio
pyrogram==1.0.1
tgcrypto
oauth2client
psycopg2-binary
sqlalchemy
httplib2
humanfriendly
pySmartDL
youtube-dl==2020.1.1
ffmpeg-python
futures
lxml
beautifulsoup4>=4.8.2,<4.8.10
psutil
pytube
myanmartools
lk21
opencv-python==4.5.2.52
Pillow
hachoir
numpy
aiohttp
